lib: libc: common: Move calloc and reallocarray into own files
This allows reusing them in different malloc implementations.
