+++ lib/libc/common/source/stdlib/calloc.c
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2018 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <stdlib.h>
+#include <string.h>
+#include <errno.h>
+#include <zephyr/sys/math_extras.h>
+
+void *calloc(size_t nmemb, size_t size)
+{
+	void *ret;
+
+	if (size_mul_overflow(nmemb, size, &size)) {
+		errno = ENOMEM;
+		return NULL;
+	}
+
+	ret = malloc(size);
+
+	if (ret != NULL) {
+		(void)memset(ret, 0, size);
+	}
+
+	return ret;
+}
