+++ lib/libc/common/source/stdlib/reallocarray.c
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2018 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <stdlib.h>
+#include <errno.h>
+#include <zephyr/sys/math_extras.h>
+
+void *reallocarray(void *ptr, size_t nmemb, size_t size)
+{
+	if (size_mul_overflow(nmemb, size, &size)) {
+		errno = ENOMEM;
+		return NULL;
+	}
+	return realloc(ptr, size);
+}
